Class providing base event functionality.

Provides methods for evaluating event occurrences. The event mechanism should only be evaluated by the simulator.

Value

Object of R6Class Event, with methods to get/set the internal parameter values and to evaluate the event mechanism.

Public fields

description

A character string giving a more descriptive overview of the event's role.

Methods

Public methods

Event$new()

Event$apply()

Event$set_parameters()

Event$get_parameters()

Event$set_id()

Event$get_id()

Event$get_name()

Event$clone()

Method new()

Create a new Event object.

Usage

Event$new(name, description, parameters, mechanism)

Arguments

name

A character string giving the name of the event. Must be unique.

description

A character string giving a more descriptive overview of the event's role.

parameters

A numeric vector giving the initial parameter values of the event.

mechanism

An expression that defines the event functionality.

Method apply()

Apply an event. Not defined for the superclass.

Usage

Event$apply()

Method set_parameters()

Set the parameter values of the event.

Usage

Event$set_parameters(parameters)

Arguments

parameters

A numeric vector giving the parameter values to set.

Method get_parameters()

Get the current parameter values.

Usage

Event$get_parameters()

Method set_id()

Assing an identifier for the event. Should not be called by the user directly.

Usage

Event$set_id(id)

Arguments

id

An integer identifier to set.

Method get_id()

Get the event identifier

Usage

Event$get_id()

Method get_name()

Get the name of the event

Usage

Event$get_name()

Method clone()

The objects of this class are cloneable with this method.

Usage

Event$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.