Class providing manipulation event functionality.

Provides manipulation event functionality. Inherited from Event.

Value

Object of R6Class ManipulationEvent.

Super class

Sima::Event -> ManipulationEvent

Methods

Public methods

ManipulationEvent$apply()

ManipulationEvent$clone()

Inherited methods

Sima::Event$get_id()

Sima::Event$get_name()

Sima::Event$get_parameters()

Sima::Event$initialize()

Sima::Event$set_id()

Sima::Event$set_parameters()

Method apply()

Apply a manipulation event

Usage

ManipulationEvent$apply(status, living)

Arguments

status

Current status of the simulated population.

living

A logical vector indicating the individuals that are still alive.

Method clone()

The objects of this class are cloneable with this method.

Usage

ManipulationEvent$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.