Class providing the simulation functionality.

Provides methods for initializing the synthetic population, parallel simulation, calibration and monitoring.

Methods

Public methods

Simulator$new()

Simulator$start_cluster()

Simulator$stop_cluster()

Simulator$run()

Simulator$reconfigure()

Simulator$configure()

Simulator$get_population()

Simulator$get_sample()

Simulator$initialize_monitor()

Simulator$clone()

Method new()

Create a new Simulator object.

Usage

Simulator$new(
  initializer = NULL,
  manipulation_events = list(),
  accumulation_events = list(),
  time = Time$new("day", Sys.Date()),
  n,
  seeds = 123
)

Arguments

initializer

A function to generate the initial population.

manipulation_events

A list of ManipulationEvent objects.

accumulation_events

A list of AccumulationEvent objects.

time

A Time object.

n

Size of the initial population.

seeds

Seed values for random number generation.

Returns

A new Simulator object.

Method start_cluster()

Start a parallel computation cluster.

Usage

Simulator$start_cluster(cl, nc, export, packages, interface = "doParallel")

Arguments

cl

A cluster object.

nc

Size of the cluster (available computing cores).

export

A character vector giving the names of objects from the global environment that should be exported to the workers.

packages

A character vector giving the names of packages that should be loaded on the workers.

interface

Interface to "foreach", must match the cluster type ("doParallel" and "doMPI" are currently supported).

Method stop_cluster()

Stop a parallel computation cluster.

Usage

Simulator$stop_cluster()

Method run()

Progress the simulation.

Usage

Simulator$run(t_sim)

Arguments

t_sim

A numeric value indicating the length of time to process.

unit

A character string giving the time unit that t_sim corresponds to.

Method reconfigure()

Configure the parameters of Event objects.

Usage

Simulator$reconfigure(p)

Arguments

p

A list of parameter values. The names of the list should match the Event object names to be configured.

Method configure()

Configure events, run a simulation and compute a summary statistic for calibration. The original status is restored after completion.

Usage

Simulator$configure(t_sim, p, output, seeds = 123, ...)

Arguments

t_sim

A numeric value indicating the length of time to process.

p

A list of parameter values. The names of the list should match the Event object names to be configured.

output

A function that computes a summary statistic from the simulation.

seeds

Seed values for random number generation.

...

Additional arguments passed to output.

Returns

The output of output_function evaluated for the resulting population.

Method get_population()

Get the status of the current population

Usage

Simulator$get_population()

Returns

A data.table of the current status

Method get_sample()

Get a sample of the current population

Usage

Simulator$get_sample(sampler, ...)

Arguments

sampler

A function that constructs the output sample.

...

Additional arguments passed to sampler.

Returns

The sample generated by sampler.

Method initialize_monitor()

Initialize monitors to track specific status variables throughout the simulation at certain intervals.

Usage

Simulator$initialize_monitor(aggregators, intervals, combiners = NULL)

Arguments

aggregators

A list of functions to compute aggregate summary statistics from the simulated population.

intervals

A numeric vector giving the time interval after which each aggregator should be evaluated.

combiners

If parallel computation is used, this is a list of functions that combine the aggregator results from each parallel worker.

Method clone()

The objects of this class are cloneable with this method.

Usage

Simulator$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.